backend/requirements.txt
Flask
flask-cors
werkzeug
pyjwt
psycopg2-binary
